gig attendance completed

Approved-by: Yogesh Sharma

+++ app/src/main/java/com/gigforce/app/modules/custom_gig_preferences/CustomPreferencesViewModel.kt
@@ -60,10 +60,15 @@ class CustomPreferencesViewModel(var owner: LifecycleOwner) : ViewModel() {
     fun markUnavaialbleTimeSlots(unavailableDataModel: UnavailableDataModel){
         if (customPreferencesDataModel != null && unavailableDataModel != null) {
             var dataModelToDelete = findDataModelForSlots(unavailableDataModel)
-            var dataModelToInsert =
-                findDataModelForUnavailableSlots(dataModelToDelete, unavailableDataModel)
-            deleteCustomPreference(dataModelToDelete)
+            if(dataModelToDelete!=null) {
+                var dataModelToInsert =
+                    findDataModelForUnavailableSlots(dataModelToDelete, unavailableDataModel)
+                deleteCustomPreference(dataModelToDelete)
             customPreferencesRepository.setData(dataModelToInsert)
+            }
+            else{
+                customPreferencesRepository.setData(unavailableDataModel)
+            }
         }
     }
